added PET_Final_tested at ABQ School of Massage
well, mixed feelings about this project.  feel like it didn't live up
to the expectations.  Certainly limited by the technology, the
pixelation the speed, the goggles, the cables, the brightness (lack of
it) in the projector.  when I'm home working on my own stuff, I
understand the state of the art, but here…well, we all have different
visions.